Moved entry logic to new file
