+++ lib/build-system/bundler.js
@@ -5,7 +5,7 @@ import path from 'path';
 export default function bundleJsModulesWithWatcher(minify, componentsByUUID) {
     const entryModulePath = path.resolve(process.cwd(), modulePath());
     build({
-        entryPoints: ['lib/reactive/watcher.js'],
+        entryPoints: ['lib/reactive/ingsoc.js'],
         inject: [entryModulePath],
         define: {
             'componentsByUUID': JSON.stringify(componentsByUUID)
