+++ lib/reactive/watcher.js
@@ -112,7 +112,6 @@ function observeNodeAttr(parentEl, selector, propName, nodeProperty, observable,
     } catch {
         return;
     }
-    console.log(nodes.length);
 
     for (const node of nodes) {
         const nodeValue = node.attributes[propName].value;
@@ -137,17 +136,4 @@ function observeNodeAttr(parentEl, selector, propName, nodeProperty, observable,
     }
 }
 
-function verifyAndRunModule(module, components) {
-    if (!Array.isArray(module)) {
-        throw new Error("Entry Module doesn't export array of components");
-    }
-    for (const [uuid, component] of Object.entries(components)) {
-        const componentNode = document.querySelector(`${component.id}[uuid="${uuid}"]`);
-        const partyMember = module.filter(value => value.id === component.id)[0];
-        const watcher = new Watcher(partyMember, true);
-
-        watcher.setup(componentNode, component.nodes);
-    }
-}
-
-verifyAndRunModule(EntryModule, componentsByUUID);
+export default Watcher;
